Backend/requirements.txt
annotated-types==0.7.0
anyio==4.12.0
asgiref==3.11.0
beautifulsoup4==4.12.3
cachetools==6.2.2
certifi==2025.11.12
charset-normalizer==3.4.4
distro==1.9.0
Django==6.0
django-cors-headers==4.9.0
django-storages==1.14.6
djangorestframework==3.16.1
dotenv==0.9.9
google-api-core==2.29.0
google-auth==2.43.0
google-cloud-core==2.5.0
google-cloud-storage==3.8.0
google-crc32c==1.8.0
google-resumable-media==2.8.0
googleapis-common-protos==1.72.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jiter==0.12.0
lxml==6.0.2
openai==2.9.0
proto-plus==1.27.0
protobuf==6.33.4
psycopg2-binary==2.9.11
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.12.5
pydantic_core==2.41.5
PyPDF2==3.0.1
python-docx==1.2.0
python-dotenv==1.2.1
requests==2.32.5
rsa==4.9.1
sniffio==1.3.1
soupsieve==2.8.1
sqlparse==0.5.4
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.1
